Prevent "upgrades" to the same version.
The recently introduced change 39c8fec13a removed a version check,
with the intent that the updater would be able to downgrade. A side
effect of removing this check is, if the advertised version is the
same as the version that's currently installed (which is the case
immediately after a successful upgrade or downgrade), the updater
gets into a loop reinstalling the same version.

This patch introduces a version check which excludes the same version
from being considered as the updater's target.